src/profiles_project/requirements.txt
Django==1.11
djangorestframework==3.6.2
pep8==1.7.1
pytz==2018.4
